Renamed interface and relocated it to common package
